Moved loadImage() from animations.PlayerAnimations to utils.ImageUtils
